requirements.txt
absl-py==0.12.0
appdirs==1.4.4
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
cycler==0.11.0
Cython==0.29.24
dataclasses==0.8
decorator==4.4.2
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphsurgeon==0.4.5
grpcio==1.34.1
h5py==2.10.0
idna==3.3
imageio==2.13.2
importlib-metadata==4.8.2
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
mock==3.0.5
networkx==2.5.1
numpy==1.19.4
nvidia-pyindex==1.0.9
oauthlib==3.1.1
onnx==1.10.2
opencv-python==4.5.4.60
opt-einsum==3.3.0
pbr==5.8.0
Pillow==8.4.0
pkg_resources==0.0.0
pkgconfig==1.5.5
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.8.1
pycuda==2020.1
pymediainfo==5.1.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytools==2021.2.9
PyWavelets==1.1.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
sk-video==1.1.10
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0+nv21.8
tensorflow-estimator==2.5.0
tensorrt==8.0.1.6
termcolor==1.1.0
testresources==2.0.1
tifffile==2020.9.3
torch
typing-extensions==3.7.4.3
uff==0.6.9
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.6.0
